Fix #8389: Charts `beginAtZero` has moved from `CartesianLinearTicks` to `CartesianLinearAxes`
