+++ docs/migrationguide/12_0_0.md
@@ -4,5 +4,8 @@
 ## Captcha
   * Is now theme aware. The default value of `theme` is now `auto` so it will attempt to detect your current theme and set it to `light` or `dark`
   
+## Charts
+  * `beginAtZero` has moved from `CartesianLinearTicks` to `CartesianLinearAxes`
+  
 ## TextEditor
   * Is now theme aware. If you are using legacy theme or no theme you may need to add CSS variables to your stylesheet to support it (https://github.com/primefaces/primefaces/issues/8064)
